refactor server routes
